+++ Makefile
@@ -32,13 +32,11 @@ SRCS                := $(SRCS_DIR)/miniRT.c \
                         $(RENDER_DIR)/render_scene.c \
                         $(RENDER_DIR)/ray_to_object.c \
                         $(RENDER_DIR)/lights.c \
-                        $(INTER_DIR)/get_closer_inter.c \
                         $(INTER_DIR)/normal_to_inter.c \
                         $(INTER_DIR)/sphere_inter.c \
                         $(INTER_DIR)/plane_inter.c \
                         $(INTER_DIR)/cylinder_inter.c \
                         $(MATHS_DIR)/vector_arithmetics.c \
-                        $(MATHS_DIR)/orthogonal_vector.c \
                         $(MATHS_DIR)/point_distance.c \
                         $(MATHS_DIR)/poly_root.c \
                         $(MATHS_DIR)/translate_point.c \
