+++ srcs/maths/vector_arithmetics.c
@@ -34,16 +34,6 @@ t_vector	add_vector(t_vector v1, t_vector v2)
 	return (result);
 }
 
-t_vector	substract_vector(t_vector v1, t_vector v2)
-{
-	t_vector	result;
-
-	result.x = v1.x - v2.x;
-	result.y = v1.y - v2.y;
-	result.z = v1.z - v2.z;
-	return (result);
-}
-
 t_vector	normalize_vector(t_vector v)
 {
 	t_vector	result;
